feat: add weighted threshold clauses and experimental nested threshold support
Most notable additions: weighted threshold clauses.
Now it's possible to have threshold like so:
["1/2", {"1/2": ["1/2", "1/2", "1/2"]}]
Arbitrary-deep weighted clauses are supported.

In addition, comes with PoC of threshold-within-threshold support:
[{"1/2": [["1"], ["1/2", "1/2", "1/2"]]}]
Note: [[]] (a logical 'and') now is not restricted to appear on
root-level.
This seems handy for multisig of multisig use-cases, when peeps
assemble into a multi-sig, their personal thresholds can be taken as
is* and prefixed with a weight
(e.g., [{"1/2": Alice's_threshold}, ...}]).
* personal threshold can be a simple threshold (e.g., "1"), but this
one can be trivially expanded, which is not the case for 'and'/multiclause as threshold.

Technical details:
1) Non-simple SignatureThreshold validation (enough_signatures) now makes use of a so-called
sig_mask: &[bool], which is an iterable of threshold's size,
each bool element of which corresponds to presence/absence of signature.

Clause(s) get their corresponding slice of sig_mask.
This is possible due to ahead-of-time derivation of:
1.clause length
2.idxes (relative to this length) corresponding to each element of clause
Which happens on construction of SignatureThreshold
(e.g., on deser, or via helpers).
This info is storred on clauses.
(ThresholdClause and ThresholdClauses)
(isn't serialized, of course)

Overall, this simplifies the outer interface, as SignatureThreshold's
enough_signature() take an arbitrary iterable of indexes, converting
it to a sig_mask that is passed on to the first clause(s) (which pass
on slices of sig_mask to their clauses and so on).

2) Various From<> helpers exist to ease construction
SignatureThreshold, previous helper-functions been ported to
From<> (interfaces preserved).

3) enough_signatures() does not error, returns bool.
Previously it errored if sig index is out of threshold length.
Now it is lenient to such out-of-size signatures, they're simply ignorred.

Final note:
Most of the code is ser/deser & tests.
ser/deser could be made prettier, I hope.